+++ backend/app/models/__init__.py
@@ -1,0 +1,30 @@
+# Re-export all models for backward compatibility
+from .chat import ChatLog
+from .storage import (
+    SafeFileSystemStorage,
+    SafeFilenameMixin,
+    SafeS3Boto3Storage,
+    get_default_storage,
+)
+from .tag import Tag, VideoTag
+from .user import User
+from .video import Video, user_directory_path
+from .video_group import VideoGroup, VideoGroupMember
+
+# Import signals to ensure they're registered
+from . import signals  # noqa: F401
+
+__all__ = [
+    "User",
+    "Video",
+    "user_directory_path",
+    "Tag",
+    "VideoTag",
+    "VideoGroup",
+    "VideoGroupMember",
+    "ChatLog",
+    "SafeFilenameMixin",
+    "SafeFileSystemStorage",
+    "SafeS3Boto3Storage",
+    "get_default_storage",
+]
